[UPDATE] Video 수정

+++ src/controllers/videoController.js
@@ -3,11 +3,10 @@ import User from "../models/User";
 // ======================================
 export const watch = async (req, res) => {
   const { id } = req.params;
-  const video = await Video.findById(id);
-  const owner = await User.findById(video.owner);
+  const video = await Video.findById(id).populate("owner");
 
   if (video) {
-   return res.render("watch", { pageTitle: video.title, video, owner});
+   return res.render("watch", { pageTitle: video.title, video });
   }
   return res.render("404", { pageTitle: "Not Found Video" });
 };
@@ -20,6 +19,9 @@ export const home = async (req, res) => {
 }
 // ======================================
 export const postEdit = async (req, res) => {
+  const {
+    user: { _id },
+  } = req.session;
   const { id } = req.params
   const video = await Video.findById(id);
   const { title, description, hashtags } = req.body;
@@ -35,16 +37,35 @@ export const postEdit = async (req, res) => {
   return res.redirect(`/videos/${id}`);
 };
 export const getEdit = async (req, res) => {
-  const { id } = req.params;
+  const { id } = req.params
+  const { 
+    user: { _id }
+  } = req.session;
   const video = await Video.findById(id);
   if (!video) {
     return res.status(404).render("404", { pageTitle: "Not Found Video" });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/")
   }
   return res.render("edit", { pageTitle :  `Edit : ${video.title}`, video});
 }
 // ======================================
 export const deleteVideo = async (req, res) => {
   const { id } = req.params;
+  const { 
+    user: { _id }
+  } = req.session;
+  const video = await Video.findById(id);
+  
+  if (!video) {
+    return res.render("404", { pageTitle: "Not Found Video" });
+  }
+
+  if (String(video.owner) !== (_id)) {
+    return res.status(403).redirect("/")
+  }
+  
   await Video.findByIdAndDelete(id);
 
   return res.redirect("/")
@@ -59,13 +80,16 @@ export const postUpload = async (req, res) => {
   const { path: fileUrl } = req.file;
   const { title, description, hashtags } = req.body;
   try { 
-      await Video.create({
+      const newVideo = await Video.create({
       title: title,
       description : description,
       fileUrl,
       owner: _id, 
       hashtags: Video.formatHashtags(hashtags),
     })
+    const user = await User.findById(_id);
+    user.videos.push(newVideo._id);
+    user.save();
     return res.redirect("/");
   } catch (error) {
       return res.status(400).render("upload",
